Core/requirements.txt
asgiref==3.3.4
certifi==2021.5.30
cffi==1.14.6
chardet==4.0.0
charset-normalizer==2.0.4
cryptography==3.4.7
defusedxml==0.7.1
Django==3.2
django-allauth==0.45.0
django-cors-headers==3.7.0
djangorestframework==3.12.4
djangorestframework-jwt==1.11.0
idna==2.10
oauthlib==3.1.1
Pillow==8.2.0
pycparser==2.20
PyJWT==1.7.1
python3-openid==3.2.0
pytz==2021.1
requests==2.25.1
requests-oauthlib==1.3.0
sqlparse==0.4.1
stripe==2.59.0
urllib3==1.26.6
